DLK-12572 Fix pdf generation for large jobs when logging is disabled
nodejs child proccess streams have a poorly documented buffer size of roughly 200KB if output on one of those streams exceeds this and there is no consumer of the stream then the process will stall and fail.

to avoid this, if we are not logging to output of wkhtmltopdf, then send all the stdout & stderr to /dev/null by telling node to ignore the output instead of sending it to the streams.